Move utility functions to config.py
